refactor: Localize rate limit error message
This commit refactors the `RateLimitException` to use a localized string resource for the error message.

Previously, the error message was a hardcoded string. Now, it fetches the translated message from `Res.string.rate_limit_exceeded`, allowing the rate limit error to be displayed in the user's language. The `RateLimitException` constructor and `safeApiCall` function have been updated to support this change.